Moved property configs inside filter config
