+++ components/common/BoardEditor/focalboard/src/components/viewHeader/filterEntry.tsx
@@ -9,7 +9,7 @@ import React, { useEffect, useMemo, useState } from 'react';
 
 import type { IPropertyTemplate, PropertyType } from 'lib/focalboard/board';
 import type { BoardView } from 'lib/focalboard/boardView';
-import type { FilterClause } from 'lib/focalboard/filterClause';
+import { propertyConfigs, type FilterClause } from 'lib/focalboard/filterClause';
 import { createFilterGroup } from 'lib/focalboard/filterGroup';
 
 import { Constants } from '../../constants';
@@ -23,134 +23,6 @@ type Props = {
   view: BoardView;
   conditionClicked: (optionId: string, filter: FilterClause) => void;
   filter: FilterClause;
-};
-
-const BooleanDataTypeConditions = ['is', 'is-not'] as const;
-
-const TextDataTypeConditions = [
-  'is',
-  'is-not',
-  'contains',
-  'does-not-contain',
-  'starts-with',
-  'ends-with',
-  'is-empty',
-  'is-not-empty'
-] as const;
-
-const NumberDataTypeConditions = [
-  'equal',
-  'not-equal',
-  'greater-than',
-  'less-than',
-  'greater-than-equal',
-  'less-than-equal',
-  'is-empty',
-  'is-not-empty'
-] as const;
-
-const DateDataTypeConditions = [
-  'is',
-  'is-not',
-  'is-before',
-  'is-on-or-before',
-  'is-on-or-after',
-  'is-between',
-  'is-empty',
-  'is-not-empty'
-] as const;
-
-const MultiSelectDataTypeConditions = ['contains', 'does-not-contain', 'is-empty', 'is-not-empty'] as const;
-
-const SelectDataTypeConditions = ['is', 'is-not', 'is-empty', 'is-not-empty'] as const;
-
-const MiscDataTypeConditions = ['is-empty', 'is-not-empty'] as const;
-
-type DataType = 'text' | 'number' | 'boolean' | 'date' | 'multi-select' | 'select' | 'misc';
-
-type DataTypeFactory<DT extends DataType, DataTypeDataTypeConditions extends readonly string[]> = {
-  datatype: DT;
-  conditions: readonly DataTypeDataTypeConditions[number][];
-};
-
-type BooleanDataTypeConfig = DataTypeFactory<'boolean', typeof BooleanDataTypeConditions>;
-type TextDataTypeConfig = DataTypeFactory<'text', typeof TextDataTypeConditions>;
-type NumberDataTypeConfig = DataTypeFactory<'number', typeof NumberDataTypeConditions>;
-type DateDataTypeConfig = DataTypeFactory<'date', typeof DateDataTypeConditions>;
-type MultiSelectDataTypeConfig = DataTypeFactory<'multi-select', typeof MultiSelectDataTypeConditions>;
-type SelectDataTypeConfig = DataTypeFactory<'select', typeof SelectDataTypeConditions>;
-type MiscDataTypeConfig = DataTypeFactory<'misc', typeof MiscDataTypeConditions>;
-
-type DataTypeConfigs =
-  | BooleanDataTypeConfig
-  | TextDataTypeConfig
-  | NumberDataTypeConfig
-  | DateDataTypeConfig
-  | MultiSelectDataTypeConfig
-  | SelectDataTypeConfig
-  | MiscDataTypeConfig;
-
-const propertyConfigs: Record<PropertyType, DataTypeConfigs> = {
-  updatedBy: {
-    datatype: 'multi-select',
-    conditions: MultiSelectDataTypeConditions
-  },
-  updatedTime: {
-    datatype: 'date',
-    conditions: DateDataTypeConditions
-  },
-  checkbox: {
-    datatype: 'boolean',
-    conditions: BooleanDataTypeConditions
-  },
-  createdBy: {
-    datatype: 'date',
-    conditions: DateDataTypeConditions
-  },
-  createdTime: {
-    datatype: 'date',
-    conditions: DateDataTypeConditions
-  },
-  date: {
-    datatype: 'date',
-    conditions: DateDataTypeConditions
-  },
-  email: {
-    datatype: 'text',
-    conditions: TextDataTypeConditions
-  },
-  file: {
-    datatype: 'misc',
-    conditions: MiscDataTypeConditions
-  },
-  multiSelect: {
-    datatype: 'multi-select',
-    conditions: MultiSelectDataTypeConditions
-  },
-  number: {
-    datatype: 'number',
-    conditions: NumberDataTypeConditions
-  },
-  person: {
-    datatype: 'text',
-    conditions: MultiSelectDataTypeConditions
-  },
-  phone: {
-    datatype: 'text',
-    conditions: TextDataTypeConditions
-  },
-  select: {
-    datatype: 'select',
-    conditions: SelectDataTypeConditions
-  },
-  text: {
-    datatype: 'text',
-    conditions: TextDataTypeConditions
-  },
-  url: {
-    datatype: 'text',
-    conditions: TextDataTypeConditions
-  }
 };
 
 function formatCondition(condition: string) {
@@ -336,17 +208,6 @@ function FilterEntry(props: Props) {
           <Typography>Delete</Typography>
         </MenuItem>
       </Menu>
-      {/* <Button
-        onClick={() => {
-          const filterGroup = createFilterGroup(view.fields.filter);
-          filterGroup.filters = filterGroup.filters.filter(
-            (o) => isAFilterGroupInstance(o) || !areFilterClausesEqual(o, filter)
-          );
-          mutator.changeViewFilter(view.id, view.fields.filter, filterGroup);
-        }}
-      >
-        <DeleteOutlinedIcon fontSize='small' />
-      </Button> */}
     </Stack>
   );
 }
